requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
Django==5.0.1
djangorestframework==3.14.0
frozenlist==1.4.1
html5lib==1.1
idna==3.6
multidict==6.0.4
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
webencodings==0.5.1
yarl==1.9.4
